folder페이지 styeld component 분리
